Create COURSEREG.php
Problem statement: https://www.codechef.com/problems/COURSEREG

Given the number of friends who want to enroll together. The capacity of the class is given. The number of people who are already enrolled is also given. Print YES, if all the friends can join the class together, if not, print NO.