Create BATH.php
Problem statement: https://www.codechef.com/problems/BATH

The capacity for a geyser is given along with the capacity for a bucket of water. If it takes 2 buckets for one person to bath with, print the total number of people that will be able to take a bath from one fully filled geyser.