Create BOOKPACK.php
Problem statement: https://www.codechef.com/problems/BOOKPACK?tab=statement

Given the number of shelves, the number of books that a shelf can hold and the number of books that a box can hold, print the number of boxes that are required. You can only put the books of one shelf in one box.