Create SINGLEUSE.php
Problem statement: https://www.codechef.com/problems/SINGLEUSE

The number of health points of a video game boss is given. The points reduced through a simple attack and a one time special attack are also given. Print the number of attacks that will be required to bring the health of the boss to 0 or below 0.