Create MELTGOLD.php
Problem statement: https://www.codechef.com/problems/MELTGOLD

Melting point of gold is given. Temperature of a kiln is given. The temperature of the kiln rises by 1 degrees / minute.  Print the number of minutes it takes for the temperature of the kiln to rise above the melting point of gold.