Create BADMINTON.php
Problem statement: https://www.codechef.com/problems/BADMINTON

Given the number of points scored by a player; the player is serving first and serves from the right side of the court; the player serves from the right side of the court when he has scored an even number of points. Print the number of times the player has served from the right side of the court.