Create CHEFCONTEST.php
Problem statement: https://www.codechef.com/problems/CHEFCONTEST

Given the time taken by Chef and Chefina to solve the problems in a contest; Also the total number of incorrect responses by them is also given. Print who won the contest considering that there is a time penalty of 10 minutes for an incorrect response.