Create ATTENDU.php
Problem statement: https://www.codechef.com/problems/ATTENDU?tab=statement

The attendance criteria for passing in a semester is given. The number of days already elapsed in the semester are given along with the number of days on which the person attended the semester. Print YES if it's still possible for the person to pass the semester and NO  otherwise.