Create LOSTWKND.php
Problem statement: https://www.codechef.com/problems/LOSTWKND?tab=statement

Given 6 integers, 5 of which represent the number of work that needs to be done on the weekdays and 6th which represents the amount by which more work need to be done. Find out if the person needs to work on the weekend or not.