Create LINCHESS.php
Problem statement: https://www.codechef.com/LP1TO204/problems/LINCHESS

Given the initial position of a person and that of N individuals. Also, the criteria for moving for the N individuals is also given. Print the number of the individual who can win the game. If one one can win print -1.