Create BRKBKS.php
Problem statement: https://www.codechef.com/problems/BRKBKS?tab=statement

Given a strength value and thickness of three bricks, calculate how many minimum number of hits it takes to break the three bricks.

Status : The problem was solved after viewing the solution description on Codechef. It said that the number of hits can be only be between 1 and 3, so each of those possible cases need to be examined separately. Using this approach resulted in a solution.