Create STONES.php
Problem statement: https://www.codechef.com/problems/STONES

Given two string, find if the characters in one string are in the 2nd string.

Status: Using the example test case, I was able to see that we need to count not just the occurrence but also the frequency of each character in the 'jewel' string. This was done using a dictionary data type without viewing the solution.