Create EQPAIR.php
Problem statement: https://www.codechef.com/problems/EQPAIR?tab=statement

Given an array of integers, find out the number of pairs of numbers where the GCD of the pair is equal to the LCM of the pair.

Status: It was realized that we need to find the nCr value for the pair of numbers. Since implementing the formula for it looked very complicated, I looked at the description of the solution, which mentioned a simplified formula of fx * (fx - 1) / 2. Using this formula for calculating the number of combinations from the frequencies of the numbers, gave the correct answer.