Create NOFIX.php
Problem statement: https://www.codechef.com/problems/NOFIX?tab=statement

Given an array of integers, find the minimum number of operations required to make all the elements of the array not equal to their position.

Status: I had to look at the solution and through trial and error I realized that it's the index that has to be increased to check if the next value in the array is equal to it.