Problem statement: https://www.codechef.com/problems/SPLITIT
Given a string, find out if there exists two strings A and B such that A
+ B = S. Where S is the given string and the string B is a subset of A.

Status: This problem was solved with the realization that if the last
character of the string appears anywhere in the rest of the string then
it would satisfy the given criteria. Solved without hints in 24 minutes.